requirements.txt
Django>=5.2.4
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
drf-spectacular>=0.28.0
cryptography>=45.0.6
psycopg2-binary>=2.9.10
django-cors-headers>=4.3.1
